Armbian SDK

+++ arch/arm/mach-sunxi/pm/pm.c
@@ -1310,7 +1310,7 @@ static int aw_pm_enter(suspend_state_t state)
 
 #ifdef CONFIG_SUNXI_ARISC
 	arisc_cpux_ready_notify();
-	arisc_query_wakeup_source((unsigned long *)(&(mem_para_info.axp_event)));
+	arisc_query_wakeup_source((unsigned int *)(&(mem_para_info.axp_event)));
 #endif
 	if (mem_para_info.axp_event & (CPUS_WAKEUP_LONG_KEY)) {
 #ifdef	CONFIG_AW_AXP
